fix(frontend): preserve preview query param when switching languages
When previewing productions or articles via ?preview=1, switching
languages in the navbar dropped the preview parameter, exiting preview
mode. Extract a shared LocaleSwitcherLinks component that appends
?preview=1 to the locale-switch href only when preview mode is active,
and use it in both SearchHeader and UnifiedHeader.
