+++ frontend/src/components/homepage/search-header/SearchHeader.tsx
@@ -2,10 +2,9 @@
 
 import { useState } from "react";
 import { Search, Menu, X } from "lucide-react";
-import { useLocale } from "next-intl";
-
 import { Link, usePathname } from "@/i18n/routing";
 import { ThemeSwitcher } from "@/components/shared/theme-switcher";
+import { LocaleSwitcherLinks } from "@/components/shared/locale-switcher-links";
 
 interface SearchHeaderProps {
     query: string;
@@ -24,7 +23,6 @@ export function SearchHeader({
 }: SearchHeaderProps) {
     const [menuOpen, setMenuOpen] = useState(false);
     const pathname = usePathname();
-    const locale = useLocale();
     const isHome = pathname === "/" || pathname === "";
     const isSearch = pathname.startsWith("/search");
     const isArticles = pathname.startsWith("/articles");
@@ -36,33 +34,7 @@ export function SearchHeader({
                 : "text-muted-foreground hover:text-foreground"
         }`;
 
-    const localeSwitcher = (
-        <span className="flex items-center gap-1 font-mono text-[10px] tracking-[1.4px] uppercase">
-            <Link
-                href={pathname}
-                locale="nl"
-                className={`transition-colors ${
-                    locale === "nl"
-                        ? "text-foreground font-medium"
-                        : "text-muted-foreground hover:text-foreground"
-                }`}
-            >
-                NL
-            </Link>
-            <span className="text-muted-foreground text-[8px]">/</span>
-            <Link
-                href={pathname}
-                locale="en"
-                className={`transition-colors ${
-                    locale === "en"
-                        ? "text-foreground font-medium"
-                        : "text-muted-foreground hover:text-foreground"
-                }`}
-            >
-                EN
-            </Link>
-        </span>
-    );
+    const localeSwitcher = <LocaleSwitcherLinks />;
 
     return (
         <header className="border-foreground border-b-2">
