+++ frontend/src/components/shared/locale-switcher-links/index.ts
@@ -1,0 +1,1 @@
+export { LocaleSwitcherLinks } from "./LocaleSwitcherLinks";
